refactor(subscribeassistantenhanced): isolate frontend project
