+++ plugins.v2/subscribeassistantenhanced/frontend/src/config/defaults.ts
