+++ plugins.v2/subscribeassistantenhanced/frontend/src/config/fields.ts
