+++ plugins.v2/subscribeassistantenhanced/frontend/src/config/i18n.ts
