+++ plugins.v2/subscribeassistantenhanced/frontend/src/config/presentation.ts
