+++ plugins.v2/subscribeassistantenhanced/frontend/src/main.ts
