+++ plugins.v2/subscribeassistantenhanced/frontend/vite.config.ts
