+++ tests/v2/subscribeassistantenhanced/frontend/values.test.ts
@@ -1,10 +1,10 @@
 import { describe, expect, it } from 'vitest'
 
-import { configDefaults } from '../../../../plugins.v2/subscribeassistantenhanced/src/config/defaults'
+import { configDefaults } from '../../../../plugins.v2/subscribeassistantenhanced/frontend/src/config/defaults'
 import {
   normalizeFiniteNumber,
   normalizeSaeConfig,
-} from '../../../../plugins.v2/subscribeassistantenhanced/src/config/values'
+} from '../../../../plugins.v2/subscribeassistantenhanced/frontend/src/config/values'
 
 describe('normalizeFiniteNumber', () => {
   it('将整数字符串转换为 number', () => {
